refactor: Add resource JSON and struct for conditional access policy
